pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "policy-gradients-pytorch"
version = "0.1.0"
description = "Simple Policy Gradient implementations in PyTorch"
authors = [
    { name = "Keishi Ishihara", email = "keishi.ishihara17@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Private :: Do Not Upload",
]
dependencies = [
    "ale-py==0.8.1",
    "gymnasium[accept-rom-license,all]==0.29.1",
    "joblib==1.4.2",
    "numpy==1.26.3",
    "pandas==2.2.3",
    "python-box[all]==7.2.0",
    "pytest==8.3.3",
    "pyyaml==6.0.2",
    "rich==13.9.2",
    "scipy==1.14.1",
    "seaborn==0.13.2",
    "tensorboard==2.18.0",
    "torch==2.3.1",
    "tqdm==4.66.5",
    "stable-baselines3==2.2.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["algos*", "tools*"]

[tool.ruff]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "third_party"]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL",  # Pylint
    "UP",  # pyupgrade
]
ignore = [
    "B905",  # Zip-without-explicit-strict
    "E501",  # Line too long
    "F403",  # Unable to detect undefined names
    "N812",  # Lowercase imported as non-lowercase
    "N999",  # Invalid module name
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments in function definition
    "PLR2004",  # Magic value used in comparison
    "PLR0915",  # Too many statements
    "PLR0402",  # Manual from import
]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
